move tests to test_drvsupport.py
